@@ -1,96 +1,83 @@
 {
-  "name": "rs-rest-service-starter",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "type": "es2020",
-  "description": "REST service starter for RS School",
-  "main": "src/server.ts",
+  "name": "trello-express-nest",
+  "version": "0.0.1",
+  "description": "",
+  "author": "",
+  "private": true,
+  "license": "UNLICENSED",
   "scripts": {
-    "start": "nodemon -L --watch '**/*.ts' --exec ts-node src/server.ts --exitcrash",
-    "start:docker": "cross-env POSTGRES_HOST=postgres nodemon -L --watch '**/*.ts' --exec ts-node src/server.ts --exitcrash",
-    "docs": "jsdoc -c jsdoc.json",
-    "lint": "eslint --ext .js,.ts --fix src/",
-    "typeorm": "ts-node ./node_modules/typeorm/cli.js --config=src/ormconfig.ts",
-    "db:create": "npm run typeorm migration:generate -- -n",
-    "db:run": "ts-node --transpile-only ./node_modules/typeorm/cli.js --config src/ormconfig.ts migration:run",
-    "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
-    "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
+    "prebuild": "rimraf dist",
+    "build": "nest build",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/ormconfig.ts",
+    "db:create": "npm run typeorm migration:generate -- -d src/migrations -n",
+    "db:run": "npm run typeorm migration:run",
+    "test": "jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
+    "test:auth": "cross-env TEST_MODE=auth jest --testMatch \"<rootDir>/test/e2e/**/*.test.js\" --noStackTrace",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+    "start": "npm run db:run && nest start",
+    "start:docker": "cross-env POSTGRES_HOST=postgres TYPEORM_MIGRATIONSRUN=true nest start",
+    "start:dev": "npm run db:run && nest start --watch",
+    "start:debug": "npm run db:run && nest start --debug --watch",
+    "start:prod": "npm run db:run && node dist/main",
+    "artillery": "artillery run artillery.yml -o result.json"
   },
-  "keywords": [
-    "rs",
-    "school",
-    "rest",
-    "node",
-    "express",
-    "autotest",
-    "starter"
-  ],
-  "license": "ISC",
-  "engines": {
-    "node": ">=14.0.0 <15"
+  "dependencies": {
+    "@nestjs/common": "^7.6.15",
+    "@nestjs/core": "^7.6.15",
+    "@nestjs/platform-express": "^7.6.15",
+    "@nestjs/platform-fastify": "^7.6.15",
+    "@nestjs/swagger": "^4.8.2",
+    "@nestjs/typeorm": "^7.1.5",
+    "@types/bcryptjs": "^2.4.2",
+    "artillery": "^1.7.6",
+    "bcryptjs": "^2.4.3",
+    "chai": "^4.3.4",
+    "cross-env": "^7.0.3",
+    "dotenv": "^10.0.0",
+    "fastify-swagger": "^4.8.2",
+    "jsonwebtoken": "^8.5.1",
+    "pg": "^8.6.0",
+    "postgres": "^1.0.2",
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^6.6.6",
+    "swagger-ui-express": "^4.1.6",
+    "typeorm": "^0.2.34",
+    "yamljs": "^0.3.0"
+  },
+  "devDependencies": {
+    "@nestjs/cli": "^7.6.0",
+    "@nestjs/schematics": "^7.3.0",
+    "@nestjs/testing": "^7.6.15",
+    "@types/express": "^4.17.11",
+    "@types/jest": "^26.0.22",
+    "@types/jsonwebtoken": "^8.5.3",
+    "@types/node": "^14.14.36",
+    "@types/supertest": "^2.0.10",
+    "@types/yamljs": "^0.2.31",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
+    "dirty-chai": "^2.0.1",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-prettier": "^3.3.1",
+    "jest": "^26.6.3",
+    "prettier": "^2.2.1",
+    "supertest": "^6.1.3",
+    "ts-jest": "^26.5.4",
+    "ts-loader": "^8.0.18",
+    "ts-node": "^9.1.1",
+    "tsconfig-paths": "^3.9.0",
+    "typescript": "^4.2.3"
   },
   "jest": {
     "testEnvironment": "node",
     "setupFilesAfterEnv": [
       "./test/setup.js"
     ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.js": [
-      "npm run lint"
-    ]
-  },
-  "dependencies": {
-    "bcrypt": "^5.0.1",
-    "cross-env": "7.0.3",
-    "dotenv": "9.0.2",
-    "express": "4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "pg": "^8.6.0",
-    "swagger-ui-express": "4.1.6",
-    "ts-node": "^10.0.0",
-    "typeorm": "^0.2.34",
-    "uuid": "8.3.2",
-    "yamljs": "0.3.0"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.11",
-    "@types/jsonwebtoken": "^8.5.2",
-    "@types/swagger-ui-express": "^4.1.2",
-    "@types/uuid": "^8.3.0",
-    "@types/yamljs": "^0.2.31",
-    "@typescript-eslint/eslint-plugin": "^4.27.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "better-docs": "^2.3.2",
-    "chai": "4.3.4",
-    "debug": "4.3.1",
-    "dirty-chai": "2.0.1",
-    "eslint": "^7.2.0",
-    "eslint-config-airbnb-base": "14.2.1",
-    "eslint-config-airbnb-typescript": "^12.3.1",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^4.0.8",
-    "husky": "4.3.8",
-    "jest": "26.6.3",
-    "jsdoc": "^3.6.7",
-    "lint-staged": "10.5.4",
-    "nodemon": "2.0.7",
-    "prettier": "2.2.1",
-    "supertest": "6.1.3"
   }
 }
