@@ -1,8 +1,6 @@
-import { join } from 'path';
 import { ConnectionOptions } from 'typeorm';
 import * as path from 'path';
-
-const dotenv = require('dotenv');
+import * as dotenv from 'dotenv';
 
 dotenv.config({
   path: path.join(__dirname, '../.env'),
@@ -14,6 +12,7 @@ const config = {
   user: process.env.POSTGRES_USER,
   password: process.env.POSTGRES_PASSWORD,
   database: process.env.POSTGRES_DB,
+  migrationsRun: process.env.TYPEORM_MIGRATIONSRUN === 'true',
 };
 
 const connectionOptions = {
@@ -23,24 +22,18 @@ const connectionOptions = {
   username: config.user,
   password: config.password,
   database: config.database,
-  migrationsRun: true,
+  migrationsRun: config.migrationsRun,
   synchronize: false,
   logging: true,
+  keepConnectionAlive: true,
   autoReconnect: true,
   reconnectTries: Number.MAX_VALUE,
   reconnectionInterval: 1000,
-  entities: [
-    join(__dirname, './resources/**/*{.ts,.js}'),
-  ],
-  migrations: [
-    join(__dirname, '../migrations/*{.ts,.js}'),
-  ],
+  entities: [path.join(__dirname, './**/*.entity{.ts,.js}')],
+  migrations: [path.join(__dirname, './migrations/*{.ts,.js}')],
   cli: {
     migrationsDir: './migrations',
   },
 } as ConnectionOptions;
 
-export const envConfig = {
-  PORT: process.env.PORT,
-};
 export default connectionOptions;
