@@ -1,10 +1,35 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { UsersModule } from './users/users.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { BoardsModule } from './boards/boards.module';
+import { TasksModule } from './tasks/tasks.module';
+import { LoginModule } from './login/login.module';
+import connectionOptions from './ormconfig';
+import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
+import { HttpErrorFilter } from './middlewares/http-error.filter';
+import { LoggingInterceptor } from './middlewares/logging.interceptor';
 
 @Module({
-  imports: [],
+  imports: [
+    TypeOrmModule.forRoot(connectionOptions),
+    UsersModule,
+    BoardsModule,
+    TasksModule,
+    LoginModule,
+  ],
   controllers: [AppController],
-  providers: [AppService],
+  providers: [
+    AppService,
+    {
+      provide: APP_FILTER,
+      useClass: HttpErrorFilter,
+    },
+    {
+      provide: APP_INTERCEPTOR,
+      useClass: LoggingInterceptor,
+    },
+  ],
 })
 export class AppModule {}
